requirements.txt
Faker==20.1.0
kafka-python==2.0.2
python-dotenv==1.0.0
